- Set safe functions dllimport to private to avoid direct usage. - Changed RegisterHotKey values to enum from uint.

+++ Desktop/Functions/ArnoldVinkInputOutput/InputOutputHotkeyRegister.cs
@@ -51,7 +51,7 @@ namespace ArnoldVinkCode
         {
             try
             {
-                bool hotkeyRegistered = AVInteropDll.RegisterHotKey(IntPtr.Zero, vHotKeyIdentifierCount, (uint)keysModifier, (uint)keysVirtual);
+                bool hotkeyRegistered = AVInteropDll.RegisterHotKey(IntPtr.Zero, vHotKeyIdentifierCount, keysModifier, keysVirtual);
                 if (hotkeyRegistered)
                 {
                     Debug.WriteLine("Registered hotkey: " + keysModifier + "/" + keysVirtual + "/ID" + vHotKeyIdentifierCount);
